+++ src/System/Collection/CollectionImmutable.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace System\Collection;
+
+class CollectionImmutable extends AbstractCollectionImmutable
+{
+  // same as perent
+}
+