+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.3] - 2021-11-07
 ### Added
-- Add feture to ```generate::class``` class, property, const and method as string
+- Add feature to ```generate::class``` class, property, const and method as string
+- Add feature to collact array ```Collection::class``` and ```CollectionImmutable::class```
 
 ### Removed
 - Remove ```Router::view```